+++ tests/geopage/models.py
@@ -16,7 +16,6 @@ from wagtail.models import Orderable, Page
 from wagtailgeowidget import geocoders
 from wagtailgeowidget.blocks import (
     GeoAddressBlock,
-    GeoBlock,
     GeoZoomBlock,
     GoogleMapsBlock,
     LeafletBlock,
@@ -227,16 +226,6 @@ class GeoStreamPage(Page):
                             ),
                         ),
                         ("map", GoogleMapsBlock(address_field="address")),
-                    ],
-                    icon="user",
-                ),
-            ),
-            (
-                "map_struct_with_deprecated_geopanel",
-                blocks.StructBlock(
-                    [
-                        ("address", blocks.CharBlock(required=True)),
-                        ("map", GeoBlock(address_field="address")),
                     ],
                     icon="user",
                 ),